Main PID algorithm programmed and tested
The main algorithm of the PID controller was programmed within the "PID"
class and tested with start parameters Kp = 0.45, Ki = 0.006 and Kd =
7.75. The system still oscilates a lot though, specially at curves. That
is because the position of the car on the lane also depends on the speed
of the car. Another PID controller will be used for the speed and the
speed setpoint will have to depend on the error between the angle
setpoint and process value.

Also the twiddle algorithm was programmed within the PID class as well.
It works well minimizing the quadratic error, but it does not reach good
results because of that dependency between the position on the lane and
the speed.