Optimized model and parameters found
Another PID controller for the speed was added, tunned and tested. Then
a cascade structure was created where the control value of the position
PID controls the setpoint of the speed PID. After the speed PID was
optimized, the position PID was optimized further until getting a good
result. The car drives well with an average speed of 50 mph.